[#46] moved remote.build to build in lib/cli
