+++ lib/cli/argv.js
@@ -4,7 +4,7 @@
 
 var path = require('path');
 var pgbuild = require('../main');
-var build = require('./remote.build');
+var build = require('./build');
 var help = require('./help');
 
 /**
