+++ lib/cli/build.js
@@ -4,7 +4,7 @@
 
 var pgbuild = require('../main');
 var console = require('./console');
-var login = require('./remote.login');
+var login = require('./login');
 
 /**
  * $ phonegap remote build <platform>
@@ -75,9 +75,8 @@ module.exports = function(argv, callback) {
         });
     };
 
-
     // call login with empty options
-    login({}, function(e, data) {
+    login({}, function(e, api) {
         console.log('login callback');
         if (e) {
             // begin prompting
@@ -88,7 +87,6 @@ module.exports = function(argv, callback) {
                 }
                 login(result, function(e,d) {
                     console.log('login with result of prompt');
-                    console.log(e, d);
                     execute(data, callback);
                });
             });
